server/requirements.txt
flask>=2.0.0
psutil>=5.8.0
pyyaml>=6.0
requests>=2.25.0